requirements_improved_pdf_converter.txt
PyMuPDF>=1.19.0
python-docx>=0.8.11
pandas>=1.3.0
openpyxl>=3.0.7
tabula-py>=2.3.0
Pillow>=9.0.0
opencv-python>=4.5.0
matplotlib>=3.5.0
camelot-py>=0.10.1
pdfplumber>=0.7.0
docxtpl>=0.16.0
ghostscript>=0.7
opencv-python-headless>=4.5
tk>=0.1.0
numpy>=1.20.0
scikit-learn>=1.0.0
scikit-image>=0.19.0
docx2pdf>=0.1.7
pytesseract>=0.3.8
